requirements.txt
openai
python-dotenv
sqlite3
streamlit
g4f
sqlite3